+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New features
 
 * text itself can be `InSteps`
+* `ListBox` helper added
 
 
 # 0.4